Allow transitions for the same Route entry to be skipped
If I have a Stack and a number of Routes under it, everything is fine.
But when some of those routes have sub-routes, we may not want animations to occur. For example, where we're handling routes as optional parts of an existing page the transition will trash any existing component state; we can solve this with redux or similar, but that adds complication where it often isn't otherwise needed.

Currently whether to do a push transition animation is mainly determined by whether the location key has changed.
This commit changes that, so that if a key has been specified on the Route inside the stack, that key is used instead.

The effect is that any sub routes of routes with keys do not animate transition.

For example:

    <Stack>
      <Route path="/" exact component={Home}/>
      <Route path="/login" component={Login}/>
      <Route path="/account" component={Account} key="account"/>
    </Stack>

    /login -> /login/complete will animate
    /login -> /account will animate
    /account -> /account/:transactionId will _not_ animate

Existing behaviour is not changed unless keys are present.